修改

+++ galaxy-mobile-sop-server/src/main/java/com/galaxyinternet/project/controller/ProjectController.java
@@ -502,14 +502,17 @@ public class ProjectController extends BaseControllerImpl<Project, ProjectBo> {
 				p.getCreateDate().trim(), "yyyy-MM-dd").getTime());
 
 		int num = projectService.updateById(project);
-		//修改项目相关的所有投资方的投资方式
-		int res=0;
-		if(project.getFinanceMode().equals("financeMode:0")){
-			res = projectService.deleteInvestById(project);
-		}else{
-			res = projectService.updateInvestById(project);
-		}
-		if (num > 0 && res>0) {
+		if (num > 0) {
+			//修改项目相关的所有投资方的投资方式
+			@SuppressWarnings("unused")
+			int res=0;
+			if(project.getFinanceMode()!=null){
+				if(project.getFinanceMode().equals("financeMode:0")){
+					res = projectService.deleteInvestById(project);
+				}else{
+					res = projectService.updateInvestById(project);
+				}
+			}
 			responseBody.setResult(new Result(Status.OK, null, "项目修改成功!"));
 			ControllerUtils.setRequestParamsForMessageTip(request,
 					project.getProjectName(), project.getId(),"2");
